+++ internal/saas/testdata/cluster/without_provider.yml
