+++ internal/saas/testdata/cluster/without_provider_and_metrics.yml
